tweaking the 'relevance' numbers, including the defaults. Now all tests pass at least, though only because we now run autodetect with a limited set of choices as many languages are very hard to unambiguously detect.
To me, string formats etc are relevant, hence we bumped up standard keyword relevance to 5, while strings and special number formats clock in at about 1. Anyway, this is never going to be better than an 'about there' kind of metric, so we call it a day now.

Meanwhile, the TSQL language spec file got fixed (relevance was ZERO, whoops!) while test/detect/checkAutoDetectResults.js has been augmented to also always list the relevance score of the actual language, so that we can see how far off it is from the first and second choice of `highlightAuto()`.